Make LSP server support `rubocop.formatAutocorrectsAll` execute command
This PR makes LSP server support `rubocop.formatAutocorrectsAll` execute command.
The user can run `rubocop -A` with the execute command in the editor.

The existing `rubocop.formatAutocorrects` execute command equivalent to `rubocop -a` can be used as follows:
https://github.com/rubocop/vscode-rubocop#manually-triggering-a-format-with-autocorrects

The new `rubocop.formatAutocorrectsAll` execute command is the provide execute command for that `rubocop -A`.