Fix projects in a subdir of HOME erroneously load ~/.rubocop.yml
It is common for projects to be in a subdirectory of $HOME, causing
a problem in the ConfigLoader where add_excludes_from_files searches
for the highest file matching .rubocop.yml and finds ~/.rubocop.yml.

This violates the documentation stating that ~/.rubocop.yml will not be
loaded when a project specific file is found first.

Adds tests for config loading behavior and extends behavior for finding
ancestor configs to make it possible to grab the highest config before
the home dir.