Add a project spec to prevent error when running `rake changelog:merge`
This PR resolves the following error when running `rake changelog:merge` in RuboCop Rails.
So it will prevent the same error from occurring in this repository.

```conosle
$ bundle exec rake changelog:merge
rake aborted!
NoMethodError: undefined method `[]' for nil
/Users/koic/src/github.com/rubocop/rubocop-rails/tasks/changelog.rb:139:in `block in contributors'
/Users/koic/src/github.com/rubocop/rubocop-rails/tasks/changelog.rb:138:in `each'
/Users/koic/src/github.com/rubocop/rubocop-rails/tasks/changelog.rb:138:in `flat_map'
/Users/koic/src/github.com/rubocop/rubocop-rails/tasks/changelog.rb:138:in `contributors'
/Users/koic/src/github.com/rubocop/rubocop-rails/tasks/changelog.rb:132:in `new_contributor_lines'
/Users/koic/src/github.com/rubocop/rubocop-rails/tasks/changelog.rb:126:in `merge_content'
/Users/koic/src/github.com/rubocop/rubocop-rails/tasks/changelog.rb:115:in `merge!'
tasks/changelog.rake:21:in `block (2 levels) in <top (required)>'
/Users/koic/.rbenv/versions/3.3.0-dev/bin/bundle:25:in `load'
/Users/koic/.rbenv/versions/3.3.0-dev/bin/bundle:25:in `<main>'
Tasks: TOP => changelog:merge
(See full trace by running task with --trace)
```

This error was caused by the absence of a space after the period:

```diff
-* [#42](https://github.com/rubocop/rubocop-rails/issues/42): Fix an error for `Rails/RedundantActiveRecordAllMethod`.([contributor_name])
+* [#42](https://github.com/rubocop/rubocop-rails/issues/42): Fix an error for `Rails/RedundantActiveRecordAllMethod`. ([contributor_name])
```

The added test verifies that there is a single space between the period and the parentheses.  And tweaks changelog entries

NOTE: For compatibility with outdated formats, if there's no contributor name, it checks that the line ends with a period.