Make `Lint/RedundantSafeNavigation` aware of constant receiver
Fixes https://github.com/rubocop/rubocop-rails/issues/1104.

This is an issue with RuboCop Rails, but it can likely be addressed as a generic issue.

So, this PR makes `Lint/RedundantSafeNavigation` aware of constant receiver.
Because use cases where a constant is `nil` are rare and an offense is detected
when the receiver is a constant. And this cop is already marked as unsafe.