Add `InternalAffairs/NodeFirstOrLastArgument` cop
This cop checks for `node.arguments.first` and `node.arguments.last`,
and corrects them to `node.first_argument` and `node.last_argument`,
respectively.

Version 1.30 of `rubocop-ast` adds support for `first_argument` on block
nodes, and is therefore required for this cop not to produce false
positives for those node types.