Fix a false positive for `Style/RedundantParentheses`
Follow up https://github.com/rubocop/rubocop/pull/12381#issuecomment-1837926968.

This PR fixes a false positive for `Style/RedundantParentheses`
when using parenthesized `lambda` or `proc` with `do`...`end` block.

`do`...`end` and `{`...`}` have different precedence, so with `do`...`end`,
parentheses are needed to correct the order of association. However, with `->`,
there is no need to worry about precedence, so the parentheses can be removed.