Allow autocorrect with display-only-fail-level-offenses
The `--display-only-fail-level-offenses` option uses
`considered_failure?` to determine what to print. The same method is
used to determine the exit code of `rubocop`, so the two options should
be entirely safe to use together.

This change is useful when using rubocop as part of a pre-commit hook.
When the hook fails becuase rubocop exited with a non-zero status, we
can use `--display-only-fail-level-offenses` to avoid having the output
cluttered up with non-blocking offenses.