Tweak the changelog URL on RubyGems.org
This PR tweaks changelog URL on RubyGems.org.
https://rubygems.org/gems/rubocop

The CHANGELOG.md is extensive, so the link of the changelog on RubyGems.org
will be updated to point to the GitHub release page, focusing on the release version.
e.g., https://github.com/rubocop/rubocop/releases/tag/v1.59.0

The idea is the same as in https://github.com/rubocop/rubocop/pull/8775.