[Fix #12574] Handle unrecognized style better in --auto-gen-config
When there are supported styles in the cop configuration, auto-generated
configuration should get an EnforcedStyle entry if the inspected code base
conforms, not to the currently select style, but to one of the other styles,
and follows that style in all places.

If the code base contains code the conforms to different styles, we should get
an Exclude entry instead. The same goes for code that doesn't conform to any of
the cop's supported styles. This last case is where we had a bug. We didn't
mark the style as unrecognized, which is why we got an EnforcedStyle instead of
an Exclude.