[Fix #12233] Make `Style/SlicingWithRange` aware of beginless range
Fixes #12233.

This PR makes `Style/SlicingWithRange` aware of redundant and beginless range.

Ruby 2.7 introduced beginless ranges. But, unlike the somewhat obscure `-1` in
`ary[1..-1]`, the `0` in `ary[0..42]` is clear as a starting point.
In fact, changing it to `ary[..42]` could potentially make it less readable.
Therefore, `ary[0..42]` should respect the original programmer's intent.
On the other hand, `ary[nil..42]` could be replaced with `ary[..42]`.
Similarly, `ary[1..nil]` could be replaced with `ary[1..]`.

Moreover, `[0..-1]` in `ary[0..-1]` is redundant and simply synonymous with `ary`.
This PR has also made adjustments to detect this case.
While this is a byproduct of supporting beginless and endless ranges,
I think including this change in `Style/SlicingWithRange` cop may be suitable
and not feel out of place.