[Fix #12179] Generate `Exclude` when `Max` is overridden
For a cop that has a Max setting, if that setting is overridden in user
configuration, creating a Max property in .rubocop_todo.yml will be
useless. User configuration will take precedence and after running
`rubocop --auto-gen-config` there can still be offenses that get reported in
the next `rubocop` run.

With this modification, we avoid the problem by generating `Exclude` instead
of `Max` in those situations.