Update the Compatibility Policy
This PR proposes to update the Compatibility Policy.

There are major version incompatibilities between Ruby 2.7 and Ruby 3.0.
And the latest stable version of Ruby on Rails, 7.1, still supports Ruby 2.7.
So, dropping support for Ruby 2.7 from RuboCop might be premature for the market at this stage.

At least for the duration that Rails 7.1 is classified as being in the "Bug Fix" phase under the [Maintenance Policy for Ruby on Rails]
(https://guides.rubyonrails.org/maintenance_policy.html), RuboCop will continue to support Ruby 2.7.

In reality, RuboCop is mostly used for code analysis with the same version of the Ruby runtime.
I think that ensuring such users can smoothly upgrade their RuboCop version justifies maintaining support
for older Ruby version 2.7 for an extended period.