[Fix #12638] Fix an `Errno::ENOENT` error when using server mode
Fixes #12638.

There is a possibility of a race condition occurring with the status file
when using server mode. The status file is updated to a new status by
`Cache.write_status_file` (`Pathname#write`), so there is no need to clear
it by deleting the file.

Due to the difficulty in reproducing this race conditions, test code for this
have not been added.