LSP mode configurable from external source
The original LSP mode determination relied on the `--lsp` command-line argument,
making it hard to set the mode from 3rd party LSP(-like) tools.
This PR refines the API to resolve this issue. Additionally, as a side effect,
it improves the mocking part of spec.

However, the APIs are still experimental and not yet mature.