Revert "Add default tags to issue templates"
Reverts #12644.

It's been a few days since the introduction, but there are false reports of the "bug" label
being applied to items that are not confirmed as bugs or are not bugs.
The default "bug" label is confusing for unconfirmed feedback.
Despite considering user likelihood to change labels, submissions are often made with the default.

For feature request, sometimes labels other than "enhancement", like "feature" or others,
are more appropriate.

Labeling can suit mechanical operations (e.g., bot), but may not fit uncertain feedback needing review,
as seen in this repository.

To avoid misleading labels on unresolved issues, the default labels will be removed from
the issue template for maintenance reasons.