Tweak an error message of `ConfigValidator`
This PR tweaks an error message of `ConfigValidator` when the config is invalid:

## Before

```console
$ bundle exec rake update_cops_documentation
(snip)

rake aborted!
RuboCop::ValidationError: Property AutoCorrect of cop Layout/EmptyComment is supposed to be
a boolean and on_command is not. (RuboCop::ValidationError)
```

## After

```console
$ bundle exec rake update_cops_documentation
(snip)

rake aborted!
RuboCop::ValidationError: Property AutoCorrect of Layout/EmptyComment cop is supposed to be
a boolean and on_command is not. (RuboCop::ValidationError)
```