Fix handling of textDocument/diagnostic
The `textDocument` parameter for the `textDocument/diagnostic` request
is of type `TextDocumentIdentifier`, which has a single property `uri`
of type `DocumentUri`.

Ref: https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_diagnostic

According to the spec the `DocumentUri` type does not hold the contents
of the document, which the handler currently relies on in order to
publish diagnostics to the client.

This breaks the LSP mode in Neovim 0.10, which now supports pull
diagnostics, and possibly other editors too. The same problem exists in
`standardrb` (https://github.com/standardrb/standard/issues/575)

I have made the `textDocument/diagnostic` hander a no-op, since
diagnostics are already being pushed to clients on
`textDocument/didChange`.