Fix an error for `Style/HashSyntax` when using `Prism::Translation::Parser`
This PR fixes the following error `Style/HashSyntax` when using `Prism::Translation::Parser` as a parser:

```console
$ bundle exec ruby -rprism/translation/parser/rubocop $(bundle exec which rubocop) \
  --only Style/HashSyntax -d lib/rubocop/cop/style/not.rb
(snip)

An error occurred while Style/HashSyntax cop was inspecting
/Users/koic/src/github.com/rubocop/rubocop/lib/rubocop/cop/style/not.rb:23:27.
undefined method `loc' for an instance of AST::Node
/Users/koic/.rbenv/versions/3.3.0/lib/ruby/gems/3.3.0/gems/parser-3.3.0.5/lib/parser/builders/default.rb:1875:
in `join_exprs'
```

Other ways to resolve this might exist, but using `RuboCop::Cop::Base#parser` is not necessary.

Currently, no repro tests have been added as it is sufficient to verify that it is not broken with the Parser gem.
And, this is for refactoring aimed at integration with `Prism`, no changelog entry will be added.

cf. https://github.com/ruby/prism/blob/v0.21.0/docs/parser_translation.md#rubocop