[Fix #12695] Handle `Include` from inherited file in parent directory
When we inherit configuration from a file in a parent or ancestor
directory, and that file is called `.rubocop.yml`, the paths in
its Include parameters are interpreted as being relative to the
file, or rather its directory. This is specific for file names
starting with `.rubocop`. The problem is that file name matching
that we do when searching for target files to inspect doesn't work
for patterns that start with `..`. Solve the problem by matching
absolute paths in these corner cases.