[Fix #12717] Re-use definition of directive comments from Style/CommentedKeyword
To make sure we don't have a different, more strict definition of what
constitutes a rubocop:disable or rubocop:todo comment in the cop compared to
elsewhere in the code, use the definition implemented in DirectiveComment.

We also get the rubocop:enable comments included in this definition, but it's
not a problem if we allow those comments in the cop, as they don't make any
sense as end-of-line comments and are unlikely to occur.