Skip `LineLength` phase on `--auto-gen-only-exclude`
Phase 1 of config auto-generation runs the `Layout/LineLength` cop on
all files to discover the `Max` line length, so it can be used by cops
which depend on it.

The `--auto-gen-only-exclude` config skips generating any `Max` configs,
meaning phase 1 is redundant, as we won't generate the `Max` for use by
phase 2.

Therefore, we can skip phase 1 entirely.