[Fix #10822] Don't cache if there are warnings
Warnings are similar to reported offenses from the users' point of view. They
expect to get the same warnings reported in each run, even if the cache is
turned on.

We add a hash for warnings per file in the global registry. If warnings have
been emitted for a certain file, we don't store the inspection results in the
cache, thus ensuring that the same warnings are printed in each run.

We were already using smart (relative) path for outputting one kind of
warning. It makes sense to use the same kind of paths consistently in the
registry.