[Fix #12731] Add cop `Lint/PossibleAssignmentToNil`
The cop detects assignments, or actually calls to setter methods,
where the receiver of the assignment can be `nil`, which would
result in an exception.

Having expressions containing `&.` to the left of the `=` never
makes sense, as far as I can see. I've tried to illustrate, in
the cop documentation examples, a couple of situations where it's
possible to make this mistake.

One is a conditional assignment `a&.value ||= 0` where the author
might have thought that they want to set `value` to `0` if `a` is
`nil`. That's not the way to do it, though.

The other is assignment in an `if` statement,
`do_something if a&.value = 0`. Here the author just forgot to
use the double equals for comparison. Normally we'd expect
`Lint/AssignmentInCondition` to handle this, but since that cop
does auto-correction, I think it's better to let this new cop
handle this particular case. It requires some analysis by the
author, and there's really no auto-correction that can be made.