Return global offenses for `Style/Copyright` when the file is empty.
Very similar to #12802

This adds a check that the returned offenses have a location that formatters can display.
While the standard formatter catched the `IndexError`, barely any others did.

Since it is not possible to autocorrect global offenses, just don't do that. Adding autocorrect support would be a bunch more work.