Remove unnecessary condition
In #10137 a fix was made that makes the condition

asgn_node.send_type? && !asgn_node.assignment_method?

unnecessary in AssignmentInCondition#skip_children?(). The spec example
'accepts = in a block followed by method call' that was added to regression
test this part of the code still passes if I remove the condition.