[Fix #12793] Fix false positives for `Style/RedundantLineContinuation`
Fixes #12793.

This PR fixes false positives for `Style/RedundantLineContinuation`
when multi-line continuations with operators.

In #12678, the use of `RuboCop::Cop::Base#parse` method was avoided due to errors in Prism.
Now, the error in Prism have been fixed:
https://github.com/rubocop/rubocop/pull/12677#issuecomment-1932189221

And given redundant line continuation patterns is too varied,
this PR have reverted to a simpler implementation using the `parse` method.