[Fix #12676] Adjust offending range in LSP
Fixes #12676.

Highlighting offenses for entire classes, modules, and methods is considered noisy in LSP.
Instead, highlighting will be applied only to the class names or method names for offenses.

To maintain compatibility for uses outside of LSP, such as with `--disable-uncorrectable`,
the behavior will be changed only when `RuboCop::LSP.enabled?` is true.