[Fix #12846] Fix an error for `RuboCop::Lockfile`
Fixes #12846.

This PR fixes an error for `RuboCop::Lockfile`
when there is no Bundler environment.

From a performance perspective as well, resolving through `require 'bundler'` is not optimal.
This PR ensures that the necessary `Bundler::LockfileParser` exists as a constant.
This approach reduces unnecessary loading caused by `require 'bundler'` and
ensures the presence of the essential `Bundler::LockfileParser`.