Add documentation and log information for enabling YJIT in LSP
This PR adds documentation on enabling YJIT for LSP client to the LSP documentation.
It also includes an update to display whether YJIT is enabled or not during LSP startup.

Although it has not been measured, due to its nature as a server, there is a potential
for improved execution efficiency with YJIT.