Respect user's intentions with `workspace/executeCommand` LSP method
Context: https://github.com/Shopify/ruby-lsp/pull/1453#issuecomment-1980179500

The `workspace/executeCommand` is an LSP method triggered by intentional user actions,
so the user's intention for autocorrection is respected.

It has not yet been implemented in RuboCop's built-in LSP, but the same could be said for `textDocument/codeAction` LSP method,
which might be implemented in the future.

This change slightly alters the behavior of `AutoCorrect: contextual` when using LSP (or the `--editor-mode` option),
and this has been documented accordingly.