Add new `consistent_either` SupportedShorthandSyntax to Style/HashSyntax
This commit adds a new `consistent_either` to `EnforcedShorthandSyntax`.

Related to feature request: https://github.com/rubocop/rubocop/issues/11436

This allows to use both explicit and shorthand syntax, similarly to
existing `either` option, but with the difference that it enforces
consistency within a single hash.

This PR extends existing `consistent` option, and skips the
`no_mixed_shorthand_syntax_check` offences when all values in the hash
are ommitable.