[Fix #12884] Allow `cop_class.documentation_url` to take a config
The current implementation is not very useful to consumers since it requires every extension to
overwrite this method.
No official extension currenty does that.

The behaviour of `--show-docs-url` is almost what's required. It takes a config to construct this url.

The url it provides falls back to the default RuboCop docs url which is undesirable in all cases I can think about. That behaviour has been removed.