Move main tests to GitHub actions
A job collects coverage, uploads the report as an artifact, and once all
matrix entries are done a separate job will download these and use
`paambaati/codeclimate-action` to pass it along to CodeClimate.
This action takes care of formatting and summing the reports.

In addition to that, I'm also fixing the spec filters which I forgot about in the other PRs