[Fix #12970] Add `CountModifierForms` option to `Metrics/BlockNesting`
Resolves #12970.

This PR adds `CountModifierForms` option to `Metrics/BlockNesting`
and set it to `false` by default.

Similar to the existing `CountBlocks`, the default is `false`.
This means that existing modifier form code will not be detected by default,
but this seems reasonable as part of the block nesting functionality.
If user require the previous behavior, user can achieve it by setting `CountBlocks` to `true`.