[Fix #12557] Make server mode aware of auto-restart for `bundle update`
Resolves #12557.

Currently, the server mode only auto-restarts due to incompatible versions of RuboCop core.
This PR will enable auto-restart when the SHA1 value of Gemfile.lock becomes incompatible.
If Gemfile.lock is missing, it will continue to refer to the RuboCop core version as before.

The advantage is that it can cover cases of version incompatibility caused by `bundle update`,
which will likely cover many scenarios like those in #12557.

The disadvantage is that the server mode may become slightly slower.
However, it is unlikely to be slow enough for users to notice. The trade-off is considered acceptable.

In the future, it might be possible to detect changes in rubocop configuration as well, but for now,
I'd like to see how things go with this change.