Display warning messages for deprecated APIs
This PR displays warning messages for deprecated APIs.

The deprecated APIs are only soft-deprecated as explicitly mentioned in the documentation.
Over three years have passed since RuboCop 1.0 was released:
https://rubygems.org/gems/rubocop/versions/1.0.0

So it seems appropriate to start displaying deprecation warnings.

This aims to prepare for RuboCop 2.0 by providing users with the opportunity to become aware of
the deprecated APIs and encouraging them to take action to avoid deprecated APIs.