Update docs to make CountAsOne values more clear.
When looking at the docs for cops that have CountAsOne, it is not immediately
clear that all support `hash` as a value.

The possible values are a little hidden and the examples do not include `hash`
in the configured values.

Example:
https://docs.rubocop.org/rubocop/cops_metrics.html#metricsclasslength

This change makes it so that the available values are shown on a new line.
It also changes the examples to have `hash` as one of the configured values.