[Fix #13232] Make server mode aware of auto-restart for local config update
Fixes #13232.

This change was also planned in #13004. It is limited to detecting local files such as `.rubocop.yml`,
but it should reduce the need to manually restart the server mode when configuration changes are detected.
Since even in LSP, detection of configuration files beyond `.rubocop.yml` is not performed,
it is expected to cover a reasonably practical use case.

Although the logic has increased compared to #13004, no significant speed degradation is anticipated.

Since it's unclear whether tracking dependencies between configuration files is practical,
it will not be handled this, at least for now.